#79 Taking a picture
Fixed the bug to go to the second view which is the picture view
