+++ HygieaWasteSortingAppIOS/HygieaWasteSortingAppIOS/LoginViewController.swift
@@ -1,0 +1,37 @@
+//
+//  LoginViewController.swift
+//  HygieaWasteSortingAppIOS
+//
+//  Created by Hesham Alghamdi on 9/16/18.
+//  Copyright © 2018 Arizona State University. All rights reserved.
+//
+
+import UIKit
+
+class LoginViewController: UIViewController {
+
+    
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
